% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_project.R
\name{quarto_project}
\alias{quarto_project}
\title{Create a custom Quarto project}
\usage{
quarto_project(name, base_dir = getwd(), default = TRUE)
}
\arguments{
\item{name}{The name of the Quarto project directory and initial \code{.qmd} file.}

\item{base_dir}{Base directory where the project should be created. Defaults to
current working directory.}

\item{default}{Set to TRUE. This will use the custom YAML within the initial
\code{.qmd} file. Change to \code{FALSE} if you would like the initial Quarto
document to have a standard YAML header. See \code{?froggeR::write_quarto()}
for more details on the default setting.}
}
\value{
A Quarto project with:\cr
* \code{.qmd}\cr
* \code{.gitignore}\cr
* \code{README.md}\cr
* \code{DATED_PROGRESS_NOTES.md}\cr
* \code{custom.scss}\cr
* \code{.Rproj}
}
\description{
This function is a wrapper for \code{quarto::quarto_create_project()} and adds some
other \code{froggeR} goodies. First, a new Quarto project is created in the provided
\code{name} directory. Then, the original \code{.qmd} file created from the
\code{quarto::quarto_create_project()} is replaced with \code{froggeR::write_quarto()}.
}
\details{
Since this will most likely be a new project directory, \code{froggeR} searches for
the existence of these other files and creates them by default:\cr
* \code{_variables.yml}\cr
* \code{.gitignore}\cr
* \code{README.md}\cr
* \code{dated_progress_notes.md}\cr
* \code{custom.scss}\cr
* \code{.Rproj}

If these files do not exist, \code{froggeR} will write them to the new project
directory.
}
\examples{
\dontrun{
# Create a temporary directory for the example
temp_dir <- tempdir()

# Create a new Quarto project with custom formatted YAML header
quarto_project('frog_project', base_dir = temp_dir, default = TRUE)

# Create a new Quarto project with generic YAML
quarto_project('frog_project_2', base_dir = temp_dir, default = FALSE)

# Clean up
unlink(file.path(temp_dir, c("frog_project", "frog_project_2")), recursive = TRUE)
}
}
