% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_brand.R
\name{save_brand}
\alias{save_brand}
\title{Save brand YAML for 'Quarto' Projects}
\usage{
save_brand(save_logos = TRUE)
}
\arguments{
\item{save_logos}{Logical. Save brand logos from `logos` directory. Default is `TRUE`.}
}
\value{
Invisibly returns `NULL` after creating system-level configuration file.
}
\description{
This function saves the current `_brand.yml` file from an existing froggeR
Quarto project. This provides a safety catch to prevent unintended overwrite if the
system-level configuration file exists.
}
\details{
The function will attempt to create a system-level configuration file from the current
froggeR Quarto project. If the system-level configuration file already exists, the
user will be prompted prior to overwrite.
}
\examples{

# Write the _brand.yml file
if (interactive()) save_brand()

}
