% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_variables.R
\name{save_variables}
\alias{save_variables}
\title{Save variables YAML for 'Quarto' Projects}
\usage{
save_variables()
}
\value{
Invisibly returns `NULL` after creating system-level configuration file.
}
\description{
This function saves the current `_variables.yml` file from an existing froggeR
Quarto project. This provides a safety catch to prevent unintended overwrite if the 
system-level configuration file exists.
}
\details{
The function will attempt to create a system-level configuration file from the current
froggeR Quarto project. If the system-level configuration file already exists, the
user will be prompted prior to overwrite.
}
\examples{

# Write the _variables.yml file
if (interactive()) save_variables()

}
