% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{Manage 'froggeR' Settings}
\usage{
settings()
}
\value{
No return value; called for side-effects.
}
\description{
This function provides useful instructions and information regarding `froggeR`
 settings.
}
\details{
This function can only run in an interactive environment. Choose to:
- Check for the existence of project- and global-level configuration files 
- Display current project-level settings
- Feedback for updating settings
- Instructions how to reuse settings across multiple `froggeR` Quarto projects.
}
\examples{
# Only run in an interactive environment
if (interactive()) froggeR::settings()

}
