% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quarto.R
\name{write_quarto}
\alias{write_quarto}
\title{Start a Quarto file with a formatted template header}
\usage{
write_quarto(
  filename = "frogs",
  path = getwd(),
  default = TRUE,
  is_project = FALSE
)
}
\arguments{
\item{filename}{The name of the file. Do not include '.qmd'}

\item{path}{The path to the main project level. Defaults to the 
current working directory.}

\item{default}{The default is set to TRUE and will create a Quarto template file
that pulls information from the folder/_variables.yml file.}

\item{is_project}{Set to \code{TRUE} when used within a Quarto project (internal use).}
}
\value{
A Quarto document with formatted YAML and two blank starter sections.
}
\description{
This function creates a new Quarto document with a pre-formatted YAML header. The 
default template requires a _variables.yml file. If none exists in the project, 
it will be created. The _variables.yml allows for a reusable custom header that 
will be applied to all \code{froggeR} default-style Quarto documents.
}
\details{
When \code{default = TRUE}, all necessary supporting files (SCSS, _variables.yml,
and _quarto.yml) will be created if they don't exist.

NOTE: Please visit \url{https://quarto.org/docs/output-formats/html-themes.html} to 
explore more Quarto theming options.
}
\examples{
\dontrun{
# To create frogs.qmd:
write_quarto(filename = "frogs", path = "path/to/project")
}
}
