% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_readme.R
\name{write_readme}
\alias{write_readme}
\title{Create a project README file}
\usage{
write_readme(path = here::here())
}
\arguments{
\item{path}{The path to the main project level. Defaults to returned value
from \code{here::here()}.}
}
\value{
A README.md template. Contains sections for project description (study
name, principle investigator, & author), project setup steps if it's to be shared
with other team members, project file descriptions, project directory descriptions,
and miscellaneous. 

NOTE: Some documentation remains to provide the user with example descriptions for 
files & directories. It is highly recommended to keep sections. This is a modifiable 
template and should be tailor-fit for your exact purpose.
}
\description{
This function is designed to first check for the existence of a README file. If none
is detected or the user chooses to overwrite the current README, a README template
is downloaded from \url{https://gist.github.com/kyleGrealis} and written at the 
project level.
}
\examples{
\dontrun{
write_readme(path = "../path_to_project")
}
}
