% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rproj.R
\name{write_rproj}
\alias{write_rproj}
\title{Create a .Rproj file}
\usage{
write_rproj(name, path = getwd())
}
\arguments{
\item{name}{The name of the project directory.}

\item{path}{The path to the main project level. Defaults to the current
working directory.}
}
\value{
A \code{.Rproj} file to initialize a RStudio project environment.
}
\description{
This function creates the \code{.Rproj} file so that any directory can be easily 
converted to a RStudio project.
}
\examples{
\dontrun{
write_rproj(name = "my_quarto_project", path = "path/to/project")
}
}
