% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scss.R
\name{write_scss}
\alias{write_scss}
\title{Create a Quarto SCSS File}
\usage{
write_scss(path = here::here())
}
\arguments{
\item{path}{Character. Path to the project directory. Default is current project
root via \code{\link[here]{here}}.}
}
\value{
Invisibly returns the path to the file.
}
\description{
Creates or opens a \code{www/custom.scss} file in a Quarto project. If the
file already exists, it is opened directly. Otherwise, the template is
downloaded from the
\href{https://github.com/kyleGrealis/frogger-templates}{frogger-templates}
repository.
}
\details{
The function creates a \code{www/custom.scss} file with styling variables,
mixins, and rules for customizing Quarto document appearance. The \code{www/}
directory is created automatically if it does not exist.
}
\examples{
\dontrun{
write_scss()
}

}
\seealso{
\code{\link{init}}, \code{\link{write_quarto}}
}
