% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yUM-package.R
\docType{package}
\name{yUM-package}
\alias{yUM}
\alias{yUM-package}
\title{yUM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
