% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ignore.R
\name{write_ignore}
\alias{write_ignore}
\title{Create an aggressive project .gitignore file}
\usage{
write_ignore(path = here::here())
}
\arguments{
\item{path}{The path to the main project level. Defaults to returned value
from \code{here::here()}.}
}
\value{
A .gitignore file. This has substanital upgrades from the basic .gitignore in
that it enhances data security. Some upgrades include ignoring R's data files (i.e.-- 
.RData, .rda, & .rds) as well as CSV, Excel, and text files. 

NOTE: Please review the contents and add other pertinent file types as appropriate.

WARNING: Check with your organization's data security team before using git with 
any sensitive and protected health information. The author assumes no liability.
}
\description{
This function is designed to first check for the existence of a .gitignore file. If none
is detected or the user chooses to overwrite the current .gitignore, a .gitignore
is downloaded from \url{https://gist.githubusercontent.com/RaymondBalise} and written 
at the project level.
}
\examples{
\dontrun{
write_ignore(path = '../path_to_project')
}
}
