% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ignore.R
\name{write_ignore}
\alias{write_ignore}
\title{Create a .gitignore File}
\usage{
write_ignore(path = here::here())
}
\arguments{
\item{path}{Character. Path to the project directory. Default is current project
root via \code{\link[here]{here}}.}
}
\value{
Invisibly returns the path to the file.
}
\description{
Creates or opens a \code{.gitignore} file in a Quarto project. If the file
already exists, it is opened directly. Otherwise, the opinionated template is
downloaded from the
\href{https://github.com/kyleGrealis/frogger-templates}{frogger-templates}
repository.
}
\examples{
\dontrun{
write_ignore()
}

}
\seealso{
\code{\link{init}}, \code{\link{write_quarto}}
}
