% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quarto.R
\name{write_quarto}
\alias{write_quarto}
\title{Start a Quarto file with a formatted template header}
\usage{
write_quarto(
  filename = "new",
  path = here::here(),
  default = TRUE,
  proj = FALSE
)
}
\arguments{
\item{filename}{The name of the file. Do not include '.qmd'}

\item{path}{The path to the main project level. Defaults to returned value
from \code{here::here()}.}

\item{default}{The default is set to TRUE and will create a Quarto template file
that pulls infomation from the folder/_variables.yml file.}

\item{proj}{Set to \code{TRUE} for Quarto projects (for internal use).}
}
\value{
A Quarto document with formatted YAML and two blank starter sections.
}
\description{
This function creates a new Quarto document with a pre-formatted YAML header. The 
default template requires a _variables.yml file. If none exists in the project, 
a prompt is provided to create the file. The _variables.yml allows for a reusable
custom header that will be applied to all \code{froggeR} default-style Quarto 
documents.
}
\details{
When \code{default = TRUE}, the project \code{path} is checked for a SCSS sheet.
If none is found, a prompt offers to create one.

NOTE: Please visit \url{https://quarto.org/docs/output-formats/html-themes.html} to 
explore more Quarto theming options.
}
\examples{
\dontrun{
# To create new_doc.qmd:
write_quarto(filename = 'new_doc', path = '../path_to_location')
}
}
