% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quarto.R
\name{write_quarto}
\alias{write_quarto}
\title{Start a Quarto file with a formatted template header}
\usage{
write_quarto(filename = "new", path = here::here(), user = "default")
}
\arguments{
\item{filename}{The name of the file. Do not include '.qmd'}

\item{path}{The path to the main project level. Defaults to returned value
from \code{here::here()}.}

\item{user}{You are the default user. You don't need to change this. 😉}
}
\value{
A Quarto document with formatted YAML and two blank starter sections.
}
\description{
This function creates a new Quarto document with a pre-formatted YAML header. The user
needs to update the Title, author details (including optional URL link, email address,
and ORCID). An abstract is included along with keywords to 'tag' the document.
}
\details{
Default options have been set for the table of contents and code output. Either set
the options to 'false' (do not quote the options in the YAML) or comment out the 
selected line(s). The default editor is source and not visual.

NOTE: The default theme for this is sandstone. Please visit
\url{https://quarto.org/docs/output-formats/html-themes.html} to explore more
theming options.
}
\examples{
\dontrun{
# To create new_doc.qmd:
write_quarto(filename = 'new_doc', path = '../path_to_location')
}
}
