% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_readme.R
\name{write_readme}
\alias{write_readme}
\title{Create a project README file}
\usage{
write_readme(path = getwd())
}
\arguments{
\item{path}{The path to the main project level. Defaults to the current
working directory.}
}
\value{
A README.md template. Contains sections for project description (study
name, principle investigator, & author), project setup steps for ease of portability,
project file descriptions, project directory descriptions, and miscellaneous. 

NOTE: Some documentation remains to provide the user with example descriptions for 
files & directories. It is highly recommended to keep these sections. This is a 
modifiable template and should be tailor-fit for your exact purpose.
}
\description{
This function is designed to first check for the existence of a README file. If none
is detected or the user chooses to overwrite the current README, a README template
is downloaded from \url{https://gist.github.com/kyleGrealis} and written at the 
project level.
}
\examples{
\dontrun{
write_readme(path = "path/to/project")
}
}
