% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_readme.R
\name{write_readme}
\alias{write_readme}
\title{Create a project README file}
\usage{
write_readme(path = getwd())
}
\arguments{
\item{path}{The path to the main project level. Defaults to the current
working directory.}
}
\value{
A README.md template. Contains sections for:\cr
* Project description (study name, principal investigator, & author)\cr
* Project setup steps for ease of portability\cr
* Project file descriptions\cr
* Project directory descriptions\cr
* Miscellaneous
}
\description{
This function is designed to first check for the existence of a README file. If none
is detected or the user chooses to overwrite the current README, a README template
is loaded from \url{https://gist.github.com/kyleGrealis} and written to the project
level.
}
\details{
This function handles the creation and/or overwriting of both \code{README.md}
and \code{dated_progress_notes.md} files. For each file, if it already exists, 
the user will be prompted whether to overwrite it. The \code{dated_progress_notes.md}
file will be initialized with the current date and a "project started" message.

NOTE: Some documentation remains to provide the user with example descriptions for
files & directories. It is highly recommended to keep these sections. However, 
this is a modifiable template and should be tailor-fit for your exact purpose.
}
\examples{
\dontrun{
write_readme(path = "path/to/project")
}
}
