% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scss.R
\name{write_scss}
\alias{write_scss}
\title{Create a Quarto SCSS file}
\usage{
write_scss(name = "custom", path = getwd())
}
\arguments{
\item{name}{The name of the scss file without extension. Default \code{name} is
"custom".}

\item{path}{The path to the main project level. Defaults to the current
working directory}
}
\value{
A \code{.scss} file to customize Quarto styling. If the file is not named
"custom.scss", the function will also attempt to update the Quarto YAML configuration.
}
\description{
This function creates the \code{.scss} file so that any Quarto project can be easily
customized with SCSS styling variables, mixins, and rules. It also updates the
Quarto YAML configuration to include the new style sheet.
}
\details{
For more information on customizing Quarto documents with SCSS, please refer to:\cr
* \url{https://quarto.org/docs/output-formats/html-themes.html#customizing-themes}\cr
* \url{https://quarto.org/docs/output-formats/html-themes-more.html}\cr
* \url{https://github.com/twbs/bootstrap/blob/main/scss/_variables.scss}
}
\examples{
\dontrun{
# Create a new temporary directory for the example
temp_dir <- tempdir()

# Create default custom.scss
write_scss(name = "custom", path = temp_dir)

# Add another style sheet
write_scss(name = "special_theme", path = temp_dir)
}
}
