% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scss.R
\name{write_scss}
\alias{write_scss}
\title{Create a Quarto SCSS file}
\usage{
write_scss(name = "custom", path = here::here())
}
\arguments{
\item{name}{The name of the scss file without extension. Default \code{name} is
"custom".}

\item{path}{The destination directory for the SCSS file. Defaults to the current 
project's base directory (ie, value of \code{here::here()}).}
}
\value{
A \code{.scss} file to customize Quarto styling. If \code{name} is not
"custom", the function will also attempt to update the Quarto document's YAML to
include the new SCSS file while preserving any existing SCSS configurations.
}
\description{
This function creates the \code{.scss} file so that any Quarto project can be easily
customized with SCSS styling variables, mixins, and rules.
}
\details{
The function includes a robust YAML handling mechanism that:
\itemize{
  \item Safely adds new project-level SCSS files without disrupting existing ones
  \item Provides instructions for adding SCSS files to Quarto YAML
}

For more information on customizing Quarto documents with SCSS, please refer to
\url{https://quarto.org/docs/output-formats/html-themes.html#customizing-themes},
\url{https://quarto.org/docs/output-formats/html-themes-more.html}, and
\url{https://github.com/twbs/bootstrap/blob/main/scss/_variables.scss} will provide
you with over 1500 lines of SCSS variables.
}
\examples{
\donttest{
# Ensure no conflicting files exist
unlink(file.path(tempdir(), "custom.scss"))

# Create the default custom.scss in a temporary directory
write_scss(name = "custom", path = tempdir())

# Ensure no conflicting files exist for another SCSS file
unlink(file.path(tempdir(), "theme_special.scss"))

# Add another SCSS file in the temporary directory
write_scss(name = "theme_special", path = tempdir())
}
}
