% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_variables.R
\name{write_variables}
\alias{write_variables}
\title{Create & write values to _variables.yml}
\usage{
write_variables(path)
}
\arguments{
\item{path}{A Quarto project name or other folder.}
}
\value{
A _variables.yml file. As described above, the contents of this file are
used in the custom Quarto header for reusability.

NOTE: Be sure to inspect the contents that are created. You can choose to add more
key:value pairs and use them in the Quarto header or anywhere else within the
Quarto document. Refer to \url{https://quarto.org/docs/authoring/variables.html} 
for more details on how to use custom variables.

WARNING: Altering the names of existing key:value pairs may cause unintended
consequences, so proceed with caution!
}
\description{
This function will create variables that are used in the
\code{froggeR::write_quarto()} custom Quarto header. Document authors can easily reuse
a consistent document structure and appearance throughout their project. The function
will look for the existence of a Quarto project \code{_variables.yml}
file. If one exists, the user will be asked to proceed to overwrite it.
}
\examples{
\dontrun{
write_variables(name = "my_quarto_project")
}
}
